refactor: extract heartbeat and PeerDown handling to mesh/heartbeat.rs
Move heartbeat loop, peer death detection, PeerDown resolution, and
MoE heartbeat policies out of mesh/mod.rs into mesh/heartbeat.rs (531 lines).

Includes:
- start_heartbeat() loop (60s gossip cycle, death detection)
- HeartbeatFailurePolicy and MoE-aware heartbeat thresholds
- resolve_peer_down, handle_peer_death, broadcast_peer_down
- broadcast_leaving (clean shutdown)
- peer_is_eligible_for_active_moe, moe_recovery_ready_at

mod.rs goes from 5,460 to 4,950 lines. All 808 tests pass.
